UPDATE: changed propertyId (cep + addressNumber) to incremental number after inserted into database; ADD: PropertyAlreadyExistsException error handler, that finds a property by cep and address number

+++ real-state/src/main/java/br/com/realstate/repository/PropertyRepository.java
@@ -16,5 +16,5 @@ public interface PropertyRepository extends MongoRepository<Property, ObjectId> 
 
     List<Property> findAllByCategory(String category);
 
-    List<Property> findAllByCep(Integer CEP);
+    List<Property> findAllByCep(Long CEP);
 }
